Fixing IGMPv3: Membership query, single join, no unicast on multicast
-- Membership Query --

Since there was no exception in the bypass driver, the membership query was
transferred by satellite. Further more, we needed an exception in the firewall
rules.

-- Single Join --

Since all multicast parameters (group addr, filtermode, source addresses, etc)
are set separately it could happen that we get first a join with an wrong
group address with is corrected by a second switch.

-- No Unicast on Multicast--

In the old implementation an unicast stream would be received even if we
have configured the device to receive multicast only.